Changed to fit Rob's pictures
STILL IN PROGRESS, DOESNT WORK PROPERLY
